Mbti를 활용한 멤버추천

+++ src/main/java/com/example/MnM/boundedContext/likeablePerson/service/LikeablePersonService.java
@@ -27,13 +27,13 @@ public class LikeablePersonService {
     }
 
     @Transactional
-    public RsData<LikeablePerson> like(Member member, String username) {
-        Optional<Member> optionalMember = memberService.findByUserName(username);
+    public RsData<LikeablePerson> like(Member member, String name) {
+        Optional<Member> optionalMember = memberService.findByName(name);
         if (optionalMember.isEmpty()) {
             return RsData.of("F-1", "존재하지 않는 유저 입니다.");
         }
 
-        if (member.getUsername().equals(username)) {
+        if (member.getName().equals(name)) {
             return RsData.of("F-2", "본인을 호감상대로 등록할 수 없습니다.");
         }
         LikeablePerson likeablePerson = LikeablePerson
@@ -48,20 +48,24 @@ public class LikeablePersonService {
         member.addFromLikeablePerson(likeablePerson);
 
         // 너를 좋아하는 호감표시 생겼어.
-        Optional<Member> memberOptional = memberRepository.findByUsername(username);
+        Optional<Member> memberOptional = memberRepository.findByName(name);
         memberOptional.get().addToLikeablePerson(likeablePerson);
 
         return RsData.of("S-1", "호감을 표시하셨습니다.", likeablePerson);
     }
+
     @Transactional
-    public RsData<LikeablePerson> cancel(LikeablePerson likeablePerson) {
+    public RsData<LikeablePerson> cancel(LikeablePerson likeablePerson, Long memberId) {
+        if (!likeablePerson.getFromMember().getId().equals(memberId)) {
+            return RsData.of("F-1", "삭제할 권한이 없습니다.");
+        }
         likeablePersonRepository.delete(likeablePerson);
 
         return RsData.of("S-1", "호감을 취소하셨습니다.", likeablePerson);
     }
 
     @Transactional
-    public RsData<LikeablePerson> modify(Member member, Long memberId, String username) {
+    public RsData<LikeablePerson> modify(Member member, Long memberId, String name) {
         if (!member.getId().equals(memberId)) {
             return RsData.of("F-1", "수정할 권한이 없습니다.");
         }
@@ -70,12 +74,12 @@ public class LikeablePersonService {
             return RsData.of("F-2", "멤버가 존재하지 않습니다.");
         }
 
-        Optional<Member> memberOptional = memberRepository.findByUsername(username);
+        Optional<Member> memberOptional = memberRepository.findByName(name);
         if (memberOptional.isEmpty()) {
             return RsData.of("F-3", "호감을 수정할 멤버가 존재하지 않습니다.");
         }
         LikeablePerson likeablePerson = likeablePersonOptional.get();
-        likeablePerson.getToMember().changeUsername(username);
+        likeablePerson.getToMember().changeUsername(name);
         return RsData.of("S-1", "호감상대를 변경하셨습니다.", likeablePerson);
     }
 }
